chore(examples): reorganize next example components
