+++ examples/next/package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "next": "^13.4.12",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.10.1"
   }
 }
