+++ examples/next/src/components/ButtonDemo.tsx
@@ -1,0 +1,13 @@
+import { Button } from "@animareflection/ui/client";
+
+import { Wrapper } from "components";
+
+const ButtonDemo = () => (
+  <Wrapper title="Button">
+    <Button w="fit-content" onClick={() => console.log("Clicked!")}>
+      Click me
+    </Button>
+  </Wrapper>
+);
+
+export default ButtonDemo;
