+++ examples/next/src/components/index.ts
@@ -1,0 +1,8 @@
+export { default as AccordionDemo } from "./AccordionDemo";
+export { default as ButtonDemo } from "./ButtonDemo";
+export { default as DrawerDemo } from "./DrawerDemo";
+export { default as IconDemo } from "./IconDemo";
+export { default as MenuDemo } from "./MenuDemo";
+export { default as ModalDemo } from "./ModalDemo";
+export { default as SliderDemo } from "./SliderDemo";
+export { default as Wrapper } from "./Wrapper";
